refactor metrics observability helpers for lint
